Use `timestamp` instead of `dateTime` which has issues on mysql v5.5
The production database being used is mysql v5.5 and it has support
issues for defaulting `dateTime` to `CURRENT_TIMESTAMP` and allowing
two timestamp values with defaults: `CURRENT_TIMESTAMP`.

This is a temporary fix and the permanent fix is to move to a
production database mysql v5.7 and revert these changes.